[cli] added printnode function and fixed a bug 1. the bug exists because in the string in atoF and atoI I didn't use '\0' as a stop 2. PrintNode() function is added and works as a charm

+++ cli/common.h
@@ -12,7 +12,7 @@
 double atoF(char* s, int len);
 int atoI(char* s, int len);
 int IsOpr(char c);
-int absF(double x);
+double absF(double x);
 
 #endif // COMMON_H
